chore(adsp-service-net-sdk,script-service): updating to dotnet framework 9.0
