+++ apps/script-service/Services/TestUtil.Spec.cs
@@ -4,7 +4,6 @@ using Moq;
 using Newtonsoft.Json;
 using Adsp.Sdk;
 using System.Text;
-using Newtonsoft.Json.Linq;
 
 namespace Adsp.Platform.ScriptService.Services;
 
@@ -95,46 +94,5 @@ internal static class TestUtil
     );
 
     return mockRestClient;
-  }
-
-  public static Dictionary<string, object>? ToDictionary(this JToken token)
-  {
-    if (token != null && token is JObject item)
-    {
-      return (Dictionary<string, object>?)DictionaryHelper(item);
-    }
-    return null;
-  }
-
-  public static object? DictionaryHelper(JToken token)
-  {
-    if (token is JObject jObject)
-    {
-      var result = new Dictionary<string, object?>();
-      foreach (var property in jObject.Properties())
-      {
-        result[property.Name] = DictionaryHelper(property.Value);
-      }
-      return result;
-    }
-    else if (token is JArray jArray)
-    {
-      var items = new List<object>();
-      foreach (var item in jArray)
-      {
-        var converted = DictionaryHelper(item);
-        if (converted is not null)
-        {
-          items.Add(converted);
-        }
-      }
-      return items;
-    }
-    else if (token is JValue jValue)
-    {
-      return jValue.Value;
-    }
-
-    return null;
   }
 }
